Refactored solutions to accept inputs as strings
