+++ 2024/day_01/task_2/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "task_2"
+name = "solution_2024_01_2"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+[lib]
+name = "solution_2024_01_2"
+path = "src/lib.rs"
