+++ 2024/day_02/task_1/Cargo.toml
@@ -1,8 +1,8 @@
 [package]
-name = "task_1"
+name = "solution_2024_02_1"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
+[lib]
+name = "solution_2024_02_1"
+path = "src/lib.rs"
