+++ 2024/day_06/task_2/Cargo.toml
@@ -1,7 +1,8 @@
 [package]
-name = "task_2"
+name = "solution_2024_06_2"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-[dependencies]
-rayon = "1.10.0"
+[lib]
+name = "solution_2024_06_2"
+path = "src/lib.rs"
