+++ 2024/day_12/task_1/Cargo.toml
@@ -1,4 +1,8 @@
 [package]
-name = "task_1"
+name = "solution_2024_12_1"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
+
+[lib]
+name = "solution_2024_12_1"
+path = "src/lib.rs"
