+++ 2024/day_13/task_1/Cargo.toml
@@ -1,6 +1,8 @@
 [package]
-name = "task_1"
+name = "solution_2024_13_1"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-[dependencies]
+[lib]
+name = "solution_2024_13_1"
+path = "src/lib.rs"
