+++ 2024/day_13/task_2/Cargo.toml
@@ -1,6 +1,8 @@
 [package]
-name = "task_2"
+name = "solution_2024_13_2"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
 
-[dependencies]
+[lib]
+name = "solution_2024_13_2"
+path = "src/lib.rs"
