+++ 2024/day_16/task_2/Cargo.toml
@@ -1,7 +1,11 @@
 [package]
-name = "task_2"
+name = "solution_2024_16_2"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
+
+[lib]
+name = "solution_2024_16_2"
+path = "src/lib.rs"
 
 [dependencies]
 aoc_utils = { path = "../../../aoc_utils" }
