+++ 2024/day_20/task_1/Cargo.toml
@@ -1,7 +1,11 @@
 [package]
-name = "task_1"
+name = "solution_2024_20_1"
 version = "0.1.0"
-edition = "2021"
+edition = "2024"
+
+[lib]
+name = "solution_2024_20_1"
+path = "src/lib.rs"
 
 [dependencies]
 aoc_utils = { path = "../../../aoc_utils" }
