+++ aoc_utils/Cargo.toml
@@ -1,6 +1,4 @@
 [package]
 name = "aoc_utils"
 version = "0.1.0"
-edition = "2021"
-
-[dependencies]
+edition = "2024"
